forgot one more bug

+++ VisualStudioProject/Src/Packets/PacketCarStatusData.cpp
@@ -31,7 +31,7 @@ void CarStatusData::fromBin(char*& data)
 	memcpy(&m_actualTyreCompound, data, sizeof(uint8_t)); data += sizeof(uint8_t);
 	memcpy(&m_visualTyreCompound, data, sizeof(uint8_t)); data += sizeof(uint8_t);
 	memcpy(&m_tyresAgeLaps, data, sizeof(uint8_t)); data += sizeof(uint8_t);
-	memcpy(&m_tyresDamage, data, sizeof(uint16_t) * 4); data += sizeof(uint16_t) * 4;
+	memcpy(&m_tyresDamage, data, sizeof(uint8_t) * 4); data += sizeof(uint8_t) * 4;
 	memcpy(&m_frontLeftWingDamage, data, sizeof(uint8_t)); data += sizeof(uint8_t);
 	memcpy(&m_frontRightWingDamage, data, sizeof(uint8_t)); data += sizeof(uint8_t);
 	memcpy(&m_rearWingDamage, data, sizeof(uint8_t)); data += sizeof(uint8_t);
